Extract transaction steps hook for ease of use
