+++ packages/ui/app/_components/popup/TransactionStepsHandler.tsx
@@ -1,5 +1,6 @@
 'use client';
 
+import { Dispatch, useReducer } from 'react';
 import { ThreeCircles } from 'react-loader-spinner';
 
 export type TransactionStep = {
@@ -11,6 +12,58 @@ export type TransactionStep = {
 export type TransactionStepsHandlerProps = {
   resetTransactionSteps: () => void;
   transactionSteps: TransactionStep[];
+};
+export type UseTransactionSteps = {
+  transactionSteps: TransactionStep[];
+  upsertTransactionStep: Dispatch<
+    | {
+        index: number;
+        transactionStep: TransactionStep;
+      }
+    | undefined
+  >;
+};
+
+export const useTransactionSteps = (): UseTransactionSteps => {
+  const [transactionSteps, upsertTransactionStep] = useReducer(
+    (
+      prevState: TransactionStep[],
+      updatedStep:
+        | { index: number; transactionStep: TransactionStep }
+        | undefined
+    ): TransactionStep[] => {
+      if (!updatedStep) {
+        return [];
+      }
+
+      const currentSteps = prevState.slice();
+
+      currentSteps[updatedStep.index] = {
+        ...currentSteps[updatedStep.index],
+        ...updatedStep.transactionStep
+      };
+
+      if (
+        updatedStep.transactionStep.error &&
+        updatedStep.index + 1 < currentSteps.length
+      ) {
+        for (let i = updatedStep.index + 1; i < currentSteps.length; i++) {
+          currentSteps[i] = {
+            ...currentSteps[i],
+            error: true
+          };
+        }
+      }
+
+      return currentSteps;
+    },
+    []
+  );
+
+  return {
+    transactionSteps,
+    upsertTransactionStep
+  };
 };
 
 function TransactionStepsHandler({
